+++ R/deployApp.R
@@ -317,7 +317,7 @@ deployApp <- function(appDir = getwd(),
   accountDetails <- accountInfo(target$account, target$server)
 
   # test for compatibility between account type and publish intent
-  if (isShinyapps(accountDetails)) {
+  if (isShinyapps(accountDetails$server)) {
     # ensure we aren't trying to publish an API to shinyapps.io; this will not
     # currently end well
     if (identical(contentCategory, "api")) {
@@ -363,7 +363,7 @@ deployApp <- function(appDir = getwd(),
                               appPrimaryDoc, assetTypeName, contentCategory, verbose, python,
                               forceRequirementsTxtEnvironment, forceGeneratePythonEnvironment)
 
-      if (isShinyapps(accountDetails)) {
+      if (isShinyapps(accountDetails$server)) {
 
         # Step 1. Create presigned URL and register pending bundle.
         bundleSize <- file.info(bundlePath)$size
@@ -483,7 +483,7 @@ getPython <- function(path) {
 
 getPythonForTarget <- function(path, accountDetails) {
   # python is enabled on Connect but not on Shinyapps
-  targetIsShinyapps <- isShinyapps(accountDetails)
+  targetIsShinyapps <- isShinyapps(accountDetails$server)
   pythonEnabled = getOption("rsconnect.python.enabled", default=!targetIsShinyapps)
   if (pythonEnabled) {
     getPython(path)